+++ src/components/HelloWorld.vue
@@ -1,3 +1,15 @@
 <template>
-    <div>Hello2</div>
+    <div>
+        <div>Hello2</div>
+    </div>
 </template>
+
+<script>
+export default {
+    data() {
+        return {
+            arr: [],
+        }
+    },
+}
+</script>