+++ src/main.ts
@@ -1,7 +1,8 @@
 import { createApp } from 'vue'
 // import App from './App'
-import App from './components/TestPromto'
+// import App from './components/TestPromto'
 // import App from './TestApp.vue'
+import App from './components/UseRoutesDemo'
 
 import './base.css'
 import './markdown.css'